requirements.txt
flask
boto3
zappa
awsudo
awscli
influxdb-client
requests
setuptools
cryptography
pyyaml
pytest
pytest-cov
